Optimized data generator for cassandra
