+++ .gitignore
@@ -1,2 +1,3 @@
 node_modules
 client/dist/bundle.js
+cassandra/*.csv
