+++ package.json
@@ -58,6 +58,7 @@
     "react-dom": "^16.8.6",
     "react-modal": "^3.8.1",
     "request": "^2.88.0",
-    "styled-components": "^4.2.0"
+    "styled-components": "^4.2.0",
+    "zlib": "^1.0.5"
   }
 }
